+++ BusinessLogic/src/com/citi/dao/ClientGoalDAO.java
@@ -1,7 +1,6 @@
 package com.citi.dao;
 
 import java.sql.Connection;
-import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -12,14 +11,14 @@ import com.citi.connections.MyConnection;
 import com.citi.pojo.ClientGoal;
 
 public class ClientGoalDAO {
-	public List<ClientGoal> fetchGoals(int clientId) {
+	public List<ClientGoal> fetchGoals(long clientId) {
 		Connection conn=MyConnection.getMyConnection();
 		String FIND_CLIENT_GOALS="SELECT GoalId,GoalAmount,CurrentYear,GoalYear,GoalsMet FROM FinancialGoals WHERE ClientId=?";
 		PreparedStatement preparedStatement;
 		List<ClientGoal> goals = new ArrayList<>();
 		try {
 			preparedStatement = conn.prepareStatement(FIND_CLIENT_GOALS);
-			preparedStatement.setInt(1,clientId);
+			preparedStatement.setLong(1,clientId);
 			ResultSet rs=preparedStatement.executeQuery();
 			while(rs.next()) { 
 				int goalId,goalYear,currentYear;
